Clean up sql and json files for logical decoding
This commit adds code to remove the sql and json files that are created
by pgcopydb when using logical decoding. This is done to avoid filling
up the disk with files that are no longer needed.

We also add a few log messages to make it easier to understand what
pgcopydb is doing.