Problem: CDC fails to replicate INSERT/UPDATE on tables with generated column
Solution: Skip generated columns while decoding logical messages. Accessing the SQLite catalog for each DML message might incur performance overhead. To avoid this, the transform process prepares an in-memory cache to hold all the tables with generated columns during the initialization of the transform. Generally, we don't expect to have many tables with generated columns, so keeping them in memory is a better trade-off and avoids performance bottlenecks.

Other possible solutions: Replacing the generated value with DEFAULT. This was not feasible because there is no way to mark keywords in the EXECUTE JSON array in the .sql file.

Related PR which skips generated column for table-data copy: #517

Signed-off-by: Arunprasad Rajkumar <ar.arunprasad@gmail.com>