Fix ordering of CFLAGS for sqlite
The CFLAGS for sqlite were being overridden by the default CFLAGS, which
caused a portion of the sqlite CFLAGS to be ignored. This commit fixes
the ordering of the CFLAGS so that the sqlite CFLAGS are applied after
the default CFLAGS.

Particularly, we need to apply the -Wno-implicit-fallthrough flag to
sqlite, as it generates a lot of warnings with the default CFLAGS on
many systems that use -Wimplicit-fallthrough=3.