Fix issues with CFLAGS
There were several issues with the CFLAGS in the pgcopydb Makefile. This patch fixes them.

1. Removal of conditional logic for USE_SECURITY_FLAGS

Historically, we needed to support some linux distros such as RHEL6 that
did not support the subset of security flags that are required for
Microsoft Compliance. However, we are now at a point where we can remove
this conditional logic and always use those security flags. This patch
removes the conditional logic and always uses the security flags.

2. Fix typo in JENKINS_CFLAGS

The JENKINS_CFLAGS variable was misspelled as CFAGS. This patch fixes
the typo.

3. Remove unnecessary override of CFLAGS

The override of CFLAGS was unnecessary and was causing the security
flags to be overridden. This patch removes the override. If we do not
have -Wformat option, -Wformat-security which is a required compiler
flag for compliance is ignored.

Relevant error message:
cc1: error: -Wformat-security ignored without -Wformat [-Werror=format-security]