Check that a table still exists before running COPY.
In Postgres pg_dump, the situation of a DROP TABLE happening concurrently to
the backup operation is prevented by taking an explicit ACCESS SHARE lock on
all the tables early.

This approach is limited to max_locks_per_transaction and can be problematic
to operate given large schemas.

In the context of pgcopydb though, we do not need to forbid DROP TABLE
happening on the source server while the migration is running, we need to be
able to skip a table that has been dropped before we get to COPY it.