Fix a deadlock situation when queueing a table's indexes.
If the index queue is full, the queue_send() operation gets a EAGAIN error
to retry later, and handles the retry policy internally. If this is done
while holding the catalog lock, then the create index processes are locked
and can't work on emptying the queue: this is a deadlock situation.

To solve it, we need to remove queue_send() from the critical section, and
the critical section needs to contain the iteration over the SQLite result
set. So first grab the result set in-memory in a new index array construct,
and then walk through that array and queue_send() on every item in there.

The index array only contains a single uint32_t value per index, so memory
size should not come to be a problem.