Introduce new options for estimating table sizes
This commit introduces a new option --estimate-table-sizes. When set,
pgcopydb will use the relpages value in system catalog to estimate the
size of each table. This is done by multiplying the number of pages by
the page size. PostgreSQL uses a default of 8KB for page sizes, but can
be changed in build time.

Optionally, the user can set the environment variable
PGCOPYDB_ESTIMATE_TABLE_SIZES to a boolean value to enable the option.

If this option is used, we run vacuumdb --analyze-only on the source to
update the relpages values before calculating the estimates.

In passing, I fixed several things including the following:
- Fix comment references to pg_autoctl
- Remove some unused functions
- Remove reference to non-existing vacuumdb command
- Remove --cache and --drop-cache references that are no longer used
- Fix typos in vacuum logs and comments
- Remove references to pgcopydb_table_size table

Calculate table size estimates on source