Add support for limiting same table concurrency (#803)
This commit adds a new option --split-max-parts to limit the number of
jobs that can be run concurrently for the same table. This option is
useful when the number of tables is large and the system is running out
of resources.

Optionally the user can also set the environment variable
COPYDB_SPLIT_MAX_PARTS to the desired value.