Refactor environment variable parsing logic (#807)
We had a repetitive pattern in the code where we checked if an environment
variable was set, and if it was, we parsed it and stored it in a target
variable.

This patch introduces a new function `get_env_using_parser` that takes
all the necesary information needed to parse the environment variable in
a struct, and does the parsing. This way, we can avoid repeating the same
code over and over again.

In passing, I also refactored our logic to set defaults before reading
environment variables, so that we can avoid repeating the same default
values in multiple places.