Set lock_timeout for replication connections
pgcopydb fails to create a replication slot on systems that have long
running transactions. This is because the CREATE_REPLICATION_SLOT
command waits for the lock to be released. This patch adds a new
connection parameter lock_timeout to the connection string for
replication connections. This parameter is set to 60 seconds by
default. This will reduce the waiting time for the CREATE_REPLICATION_SLOT
command to fail.