Fix out-of-memory issue in creating our pg_restore list file.
Instead of creating the whole contents of the file in-memory, write it to
disk one line at a time, keeping only that line in-memory during the whole
operation.

Also arrange to code to only read the pg_restore --list output one line at a
time, no context is needed to be able to parse this file format (pg_dump
archive Table Of Contents).