Use our own sqlite3 binary in testing environments.
The OS sqlite3 binary often is a different version than the one we have
built pgcopydb with, and also might use different build-time options (such
as support for the JSON API or other elements).

Ensure that we're using the sqlite3 binary that matches with the version of
sqlite selected at pgcopydb build-time.