Use compose plugin of docker (#859)
`docker-compose` is an external tool written in python and is deprecated
for a while. Github actions also removed support for that recently.

docker cli has a plugin named compose that is available as a command of
the cli as in `docker compose`.

`docker-compose` and `docker compose` do not have same specifcations and
it is not always a trivial task to migrate from one to another. This
commit will only change all the references to the `docker-compose` to
`docker compose` in our codebase.