Use exec form of Dockerfile CMD instructions
This change updates the Dockerfile CMD instructions to use the exec form
instead of the shell form. This is a best practice for Dockerfiles as it
avoids the overhead of running a shell process.

One downside of shell form is that it does not pass signals to the
process. This can cause the process to not shut down cleanly when
stopping the container. The exec form does not have this issue.