Generate optstrings automatically
This commit introduces a new function construct_optstring that
constructs the optstring for getopt_long based on the long_options array.

We used to manually edit the optstring in each command's getopts
function. This was error-prone and tedious. This commit removes the
manual editing and replaces it with a call to construct_optstring.