test_decoding transform: skip toast columns with unchanged values (#877)
test_decoding plugin may not include the value of a TOAST column if it
has not changed. In that case, instead of the value of the column the
string "unchanged-toast-datum" is included in the output. In that case,
we should remove the column from the range table entry list in the
generated SQL queries instead of including it and setting that column to
the value "unchanged-toast-datum".

There is one strange case that if a table has only toast columns, and an
update query that does not update any value is executed, we end up
filtering all the range table entries and skip executing any sql queries
as it will be no-op.