Fix docker compose external volume (#860)
* Fix docker-compose volume sharing between services.

In unit tests where we use both an inject and a test services, we need the
inject service to access the same pgcopydb work directory as the test
service. For that we use an external docker volume.

The previous implementation of that idea was found to be wrong, this is
fixing it by properly using docker APIs for shared "external" volumes.

* Attempt to fix github action errors.

On my local environment docker volume create turned out to not be necessary,
which is good news because we can (and probably want to) keep the volumes
around between test rounds. A cleanup script has even been implemented to
restart fresh.

On GitHub actions, it looks like we might need to first create the docker
volume to then be able to use it. The error message is not very helpful (it
just says: docker-compose: no such file or directory), but it's worth trying
that.

* Add `make attach` to tests using a docker volume.

This runs an interactive docker container (bash) with the volume mounted for
exploration / debug / inspect.

* Refrain from re-assigning the PATH variable in Makefiles.

* The make fix-volume operation does not need an interactive terminal.

* Fix `docker compose version` command.