Fix docker-compose volume sharing between services.
In unit tests where we use both an inject and a test services, we need the
inject service to access the same pgcopydb work directory as the test
service. For that we use an external docker volume.

The previous implementation of that idea was found to be wrong, this is
fixing it by properly using docker APIs for shared "external" volumes.